requirements.txt
google-genai>=1.0.0
pypdf2>=3.0.0
pdfplumber>=0.10.0
python-docx>=0.8.11
pyyaml>=6.0
python-dotenv>=1.0.0
flask>=3.0.0
werkzeug>=3.0.0
